requirements.txt
requests
PyJWT==2.8.0
httpx==0.27.2
cryptography